+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # ChangeLog
+## 2.0.3 (2024-10-23)
+- Added Export/Import Dialog
+
 ## 2.0.2 (2024-10-16)
 - Added JSON debug function
 - Added german GUI translation